requirements.txt
langchain-core
langchain-google-genai
langgraph
langgraph-checkpoint-redis
mcp
langchain-mcp-adapters
langsmith
fastapi>=0.116
uvicorn
python-dotenv
requests
jax
flax
sentencepiece
tensorflow
numpy
Pillow
einops
overrides
ml_collections
kagglehub

# Benchmark dependencies
beautifulsoup4>=4.11.0
lxml>=4.9.0
google-genai>=1.0.0
transformers  # CLIP baseline
torch